Initial version using long for everything

+++ src/main/java/ipsis/woot/farming/IRecipeProgressTracker.java
@@ -10,8 +10,8 @@ public interface IRecipeProgressTracker {
     void reset();
     void setPowerRecipe(PowerRecipe powerRecipe);
     void setPowerStation(IPowerStation powerStation);
-    int getConsumedPower();
+    long getConsumedPower();
 
     // For restore on load only
-    void setConsumedPower(int power);
+    void setConsumedPower(long power);
 }
